+++ src/combat/battlers/battler.gd
@@ -8,9 +8,11 @@
 @tool
 class_name Battler extends Node2D
 
-
-## Emitted when the battler finished their action and arrived back at their rest position.
-signal action_finished
+## Emitted whenever the Battler caches a valid action.
+signal action_cached
+## Emitted whenever the Battler's turn is finished. Thos should emit only after all actions and 
+## animations are complete.
+signal turn_finished
 ## Forwarded from the receiving of [signal BettlerStats.health_depleted].
 signal health_depleted
 ## Emitted when taking damage or being healed from a [BattlerHit].
@@ -19,10 +21,6 @@ signal health_depleted
 signal hit_received(value: int)
 ## Emitted whenever a hit targeting this battler misses.
 signal hit_missed
-## Emitted when the battler's `_readiness` changes.
-signal readiness_changed(new_value)
-## Emitted when the battler is ready to take a turn.
-signal ready_to_act
 ## Emitted when modifying `is_selected`. The user interface will react to this for
 ## player-controlled battlers.
 signal selection_toggled(value: bool)
@@ -34,6 +32,7 @@ const GROUP: = "_COMBAT_BATTLER_GROUP"
 @export var stats: BattlerStats = null
 ## Each action's data stored in this array represents an action the battler can perform.
 ## These can be anything: attacks, healing spells, etc.
+
 @export var actions: Array[BattlerAction]
 
 ## Each Battler is shown on the screen by a [BattlerAnim] object. The object is created dynamically
@@ -91,24 +90,10 @@ const GROUP: = "_COMBAT_BATTLER_GROUP"
 				ai = new_instance
 				add_child(ai)
 
-## The [CombatActor] object that will determine the Battler's combat behavior. A Battler without an CombatActor
-## is just a dummy object that will not take turns or perform actions.
-@export var actor_scene: PackedScene:
-	set(value):
-		if value == actor_scene:
-			return
-
-		actor_scene = value
-
-		if not is_inside_tree():
-			await ready
-
-		if actor:
-			actor.queue_free()
-			actor = null
-
-#TODO: Battler.is_player is redundant. Use that defined by CombatActor instead.
-## Player battlers are controlled by the player.
+## If this is [b]true[/b], this actor is controlled by the player. Use this to
+## differentiate between player-controlled actors and AI-controlled ones.
+## [/n][/n] Note that player Battlers may be controlled by an AI controller, but enemy Battlers may
+## not be controlled by the player.
 @export var is_player: = false:
 	set(value):
 		is_player = value
@@ -116,28 +101,18 @@ const GROUP: = "_COMBAT_BATTLER_GROUP"
 			var facing: = BattlerAnim.Direction.LEFT if is_player else BattlerAnim.Direction.RIGHT
 			anim.direction = facing
 
-## Reference to the Battler's child [CombatActor], which controls it's combat behaviour.
-## Note that this value is assigned automatically with reference to [member actor_scene].
-var actor: CombatActor = null
-
 ## Reference to this Battler's child [CombatAI] node, if applicable.
 var ai: CombatAI = null
 
 ## Reference to this Battler's child [BattlerAnim] node.
 var anim: BattlerAnim = null
 
-# TODO: this property belongs with the CombatActor.
-## If `false`, the battler will not be able to act.
+## If this is [b]true[/b], this Battler can take turns.
 var is_active: bool = true:
 	set(value):
 		is_active = value
 
 		set_process(is_active)
-
-## The turn queue will change this property when another battler is acting.
-var time_scale := 1.0:
-	set(value):
-		time_scale = value
 
 ## If `true`, the battler is selected, which makes it move forward.
 var is_selected: bool = false:
@@ -155,25 +130,22 @@ var is_selectable: bool = true:
 		if not is_selectable:
 			is_selected = false
 
-## When this value reaches `100.0`, the battler is ready to take their turn.
-var readiness := 0.0:
+## Combat happens in two phases. In the first phase each Battler - player and enemy - select an
+## action. In the second phase that action is carried out.
+## When selecting actions, a Battler caches it in the following memeber. The cached action is set
+## to null after execution.
+var cached_action: BattlerAction = null:
 	set(value):
-		readiness = value
-		readiness_changed.emit(readiness)
+		cached_action = value
+		action_cached.emit()
 
-		if readiness >= 100.0:
-			readiness = 100.0
-			stats.energy += 1
 
-			ready_to_act.emit()
-			set_process(false)
+static func sort(a: Battler, b: Battler) -> bool:
+	return a.stats.speed > b.stats.speed
 
 
 func _ready() -> void:
-	if Engine.is_editor_hint():
-		set_process(false)
-
-	else:
+	if not Engine.is_editor_hint():
 		assert(stats, "Battler %s does not have stats assigned!" % name)
 
 		add_to_group(GROUP)
@@ -187,27 +159,33 @@ func _ready() -> void:
 			is_selectable = false
 			health_depleted.emit()
 		)
+		
+		# Similarly, duplicate all actions since some may be shared across multiple Battlers and
+		# each needs to point to the action source (i.e. this Battler).
+		var battler_roster: = _get_roster()
+		assert(battler_roster != null, "%s is not a descendent of the BattlerRoster!" % name)
+		
+		var duplicate_actions: Array[BattlerAction] = []
+		for battler_action in actions:
+			var duplicate_action: = battler_action.duplicate()
+			duplicate_action.source = self
+			duplicate_action.battler_roster = battler_roster
+			duplicate_actions.append(duplicate_action)
+		actions = duplicate_actions
 
 
-func _process(delta: float) -> void:
-	readiness += stats.speed * delta * time_scale
+## [method BattlerAction.execute] the Battler's [member cached_action].
+func act() -> void:
+	if cached_action:
+		stats.energy -= cached_action.energy_cost
 
-
-func act(action: BattlerAction, targets: Array[Battler] = []) -> void:
-	set_process(false)
-
-	stats.energy -= action.energy_cost
-
-	# action.execute() almost certainly is a coroutine.
-	@warning_ignore("redundant_await")
-	await action.execute(self, targets)
-	if stats.health > 0:
-		readiness = action.readiness_saved
-
-		if is_active:
-			set_process(true)
-
-	action_finished.emit.call_deferred()
+		# cached_action.execute() is almost certainly is a coroutine.
+		@warning_ignore("redundant_await")
+		await cached_action.execute()
+	
+	# Flag that the Battler has acted by resetting the cached action.
+	cached_action = null
+	turn_finished.emit.call_deferred()
 
 
 func take_hit(hit: BattlerHit) -> void:
@@ -218,5 +196,21 @@ func take_hit(hit: BattlerHit) -> void:
 		hit_missed.emit()
 
 
-func is_ready_to_act() -> bool:
-	return readiness >= 100.0
+# Iteratively search this node's parents for the BattlerRoster. Battler's must be descendants of the
+# BattlerRoster. Used to setup this Battler's BattlerActions.
+func _get_roster() -> BattlerRoster:
+	var parent: = get_parent()
+	while parent != null:
+		if parent is BattlerRoster:
+			return parent as BattlerRoster
+		parent = get_parent()
+	return null
+
+
+func _to_string() -> String:
+	var msg: = "%s (Battler)" % name
+	if not is_active:
+		msg += " - INACTIVE"
+	elif cached_action == null:
+		msg += " - HAS ACTED"
+	return msg