+++ src/combat/ui/effect_labels/ui_effect_label_builder.gd
@@ -8,8 +8,8 @@ class_name UIEffectLabelBuilder extends Node2D
 @export var missed_label_scene: PackedScene
 
 
-func setup(battler_data: BattlerList) -> void:
-	for battler in battler_data.get_all_battlers():
+func setup(battler_data: BattlerRoster) -> void:
+	for battler in battler_data.get_battlers():
 
 		battler.hit_missed.connect(func _on_battler_hit_missed() -> void:
 			var label: = missed_label_scene.instantiate()